Add a few helpful operators
In practice, there aren't many reasons to send more than a single value across each "edit session", and using a MutableProperty is less awkward than a signal/observer pair.

This changeset introduces a pair of operators that allow a Signal{,Producer} to be directed into a MutableProperty that holds the current EditSession for an Editor.